fix: [#54858] Fix event generator retrying logic
From the logs, we see that both integrations start failing as soon as
the production deployment comes online. The hypothesis is, that because
the request is retried after 5 seconds in case of a failure, the Omada
API server is put under constant high load as soon as one integration
enters a failure loop. When the second integration, somewhere in the
next 30 minutes, has to generate events, it also fails, and the server
is now fully loaded trying to generate a response for both integrations
at once.